[runtime,storage] Simplify the deferred-dir-sync implementation
From a simplification review of the deferral commits: collapse the
duplicated plain-write-plus-full-sync slow path in the tokio
write_at_sync (a statement-level cfg gate with an early-returning
fast path leaves the slow path stated once, matching iouring's
shape), share the deferral contract assertions between the tokio and
iouring tests via a storage::tests helper (which also gains iouring
the reopened-handle case), reuse Writable::sync_from for the fresh
tail sync in clear, and state each durability fact once at its owning
site (open_versioned owns the contract; sync, write_at_sync, and
DirSync link to it, with the empty-write early return now called
out).
