+++ storage/src/journal/contiguous/blobs.rs
@@ -375,15 +375,13 @@ impl<E: Context> Writable<E> {
         }
         let _ = self.metrics.tracked.try_set(0);
         self.tail = self.partition.open(tail_blob).await?;
-        // Make the fresh tail durable (creation defers that to the first sync, including its
-        // directory entries) before callers durably record state that references it.
-        self.tail.sync().await.map_err(Error::Runtime)?;
-        self.metrics.synced.inc();
         self.metrics.tracked.inc();
         self.oldest_blob_index = tail_blob;
         self.sealed.clear();
         self.sealed_snapshot = None;
-        Ok(())
+        // Make the fresh tail durable (creation defers that to the first sync, including its
+        // directory entries) before callers durably record state that references it.
+        self.sync_from(tail_blob).await
     }
 
     /// Make every blob from `start_blob` onward durable.
